requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.7.0
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
django-htmx==1.21.0
django-js-asset==2.2.0
django-template-partials==24.4
django-timezone-field==7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.55.3
idna==3.10
isodate==0.7.2
kombu==5.4.2
lxml==5.3.0
Markdown==3.7
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
zeep==4.3.1
zopfli==0.2.3.post1